http: Allow configuration of the URI from /new
This patch adds a new flag, `--host`, which permits runtime
configuration of the URI returned by the `/new` handler. The default
behavior is the current hardcoded behavior, which always directs clients
to https://discovery.etcd.io.

Example usage:

    bin/discovery.etcd.io --host=https://discovery.example.com

The given string will be used verbatim and is not parsed. It is
completely possible to return malformed URIs if not careful.

Future work should perhaps react to the Host header, but further study
is needed regarding the ELB CoreOS uses in production; at the very
least, TLS should be preferred, so the Host header might not always be
the best default for this purpose.

Fixes: #3